feat(LadderGame) : WinningResult 생성 역할 분배를 위한 LadderGame 클래스 생성
